Add files via upload
The purpose of this app will be for anyone to be able to see what their daily weight is and how it fluctuates. You can use it to track weight loss, gain, or maintain weight. You can set a goal weight higher or lower than your current weight, and your progress will update when you add your daily weight to the application. The components that need to be included will be a login page, a home screen, navigational buttons at the bottom for quick entry of daily weight, a line graph showcasing weight fluctuation, and notifications to enter weight and where their progress is. The line graph will showcase how much max weight they should be losing weekly based on scientific feedback pulled from multiple sources online and A simplified version of my weight-tracking app description will benefit me and the customer the most. A straightforward description of the app and its functionality will ensure users don’t get frustrated after downloading if it’s missing a function they were looking for.  The description will include features, the app's basic layout, and its primary function and use. The icon that would be best to use would be a simple scale. This would let people know it's for some weighing function but wouldn’t confuse users on the icon or why it’s set up that way. The components that need to be included will be a login page, a home screen, navigational buttons at the bottom for quick entry of daily weight, a line graph showcasing weight fluctuation, and notifications to enter weight and where their progress is. The line graph will showcase how much max weight they should be losing weekly based on scientific feedback pulled from multiple sources online and their current weight loss or gain in different colors so they are easily distinguishable. I believe some users will want to watch their weight and be cautious of where they are and the normal fluctuations throughout the year. They will choose this simple app because hundreds out there focus mainly on weight loss. Hence, they want all this information and recommend recipes and how to lose weight, which can be annoying to someone who wants to be aware of their body but doesn’t want to change something so simple will be better for them. 
Screens and features that will be needed to produce a user-centered UI will be the login screen right when you open the application, a button or insert to enter in daily weight and time the weight was taken, a questionnaire when setting up the app that can be edited to input goal weight, whether you are increasing, maintaining or losing weight. 
The functional app will require input strings for the username and password when logging in and will crossreference the input data into the memory bank or storage created to pull the login information to ensure the credentials are correct. Once the information is validated and accurate, the user can move to the home screen. When the user hits the button for daily weight, a box will open for the weight to be put in. There will be tabs at the bottom of the page to take them to different screens based on what they would like to do. They can navigate to the line graph to see all weights entered against the goal weight on the same graph, just different lines with different colors.